Replace AdvancedDLSupport with P/Invoke

+++ build-librandomx.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 set -e
 
+if [[ "$CI" = "true" ]]; then
+  set -vx
+fi
+
 if ! command -v git > /dev/null; then
   echo "No git in the system; install git first." > /dev/stderr
   exit 1
@@ -20,9 +24,12 @@ Darwin)
   ;;
 esac
 destination="$(dirname "$0")/native/$shared_lib_filename"
-if [[ -f "$destination" ]]; then
+if [[ -f "$destination" && "$CI" != "true" ]]; then
   echo "$destination already exists; this script will do nothing." > /dev/stderr
   exit 0
+fi
+if [[ ! -d "$(dirname "$0")/native" ]]; then
+  mkdir -p "$(dirname "$0")/native"
 fi
 
 tmpdir="$(mktemp -d)"
@@ -37,6 +44,13 @@ git clone \
   git://github.com/tevador/RandomX.git \
   "$tmpdir"
 pushd "$tmpdir"
+if [[ "$(uname -s)" = Linux ]] && command -v gcc > /dev/null; then
+  export CC=gcc
+  export CXX=g++
+elif [[ "$(uname -s)" = Darwin ]] && command -v clang > /dev/null; then
+  export CC=clang
+  export CXX=clang++
+fi
 cmake -DARCH=native -DBUILD_SHARED_LIBS=ON
 make
 popd
